rename heicConverter
