+++ src/services/heicConverter.ts
